requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
entrypoints==0.4
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.4.2
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.23.0
postgrest==0.16.11
protobuf==5.27.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
realtime==2.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
storage3==0.7.7
streamlit==1.37.1
streamlit-aggrid==1.0.5
streamlit-image-zoom==0.0.4
streamlit-javascript==0.1.5
streamlit_folium==0.22.0
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tenacity==8.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
xyzservices==2024.9.0
yarl==1.11.1
